Classes to implement Caching for cover-arts
